requirements.txt
Django>=3.2
django-simple-task
petl
python-dateutil
requests
uvicorn
